requirements.txt
python-telegram-bot[job-queue]==20.7
groq==0.9.0
python-dotenv==1.0.1
aiosqlite==0.20.0
aiohttp==3.10.0
